完成msvc的兼容，后续在windows下用msvc编译，mingw对异常处理兼容不是很好

+++ src/common/src/StringUtils.cpp
@@ -150,7 +150,7 @@ namespace OneCommon
 				start = i + 1;
 			}
 		}
-		if (ignoreBlank == false || start < size)
+		if (ignoreBlank == false || start < (int)size)
 		{
 			lines.push_back(str.substr(start, size - start));
 		}
